added 'make install' to CTF C++ build system. Changed setup_wrapper.py to setup.py. Turned on MPI auto-initialization upon CTF import. Took out MPI_start/end calls from tests.
Took out build targets for python tests. New build process for python version is simple

./configure
make
sudo make install
pip install .

then can run e.g. python test/python/test_einsum.py or start a jupyter notebook and run 'import ctf', etc.

+++ src_python/ctf_ext.cxx
@@ -102,7 +102,7 @@ namespace CTF_int{
 	template void tensor::exp_helper<std::complex<double>, std::complex<double>>(tensor* A);
   // exp_helper when casting == unsafe
 	template void tensor::exp_helper<int64_t, float>(tensor* A);
-	template void tensor::exp_helper<int64_t, double>(tensor* A);
+	template void tensor::exp_helper<int32_t, float>(tensor* A);
 
 	// ctf.all() function in c++ file (add more type)
 	template void all_helper<double>(tensor * A, tensor * B_bool, char const * idx_A, char const * idx_B);
